requirements.txt
fastmcp>=2.0.0
json-logic-py>=0.7.0
pydantic>=2.0.0
typing-extensions>=4.0.0 